revisi fix

+++ BusTicketActivity/app/src/main/java/com/example/busticketactivity/tiketmenu/DetailTiketActivity.kt
@@ -8,18 +8,30 @@ import androidmads.library.qrgenearator.QRGContents
 import androidmads.library.qrgenearator.QRGEncoder
 import androidx.appcompat.app.AppCompatActivity
 import com.example.busticketactivity.R
+import com.example.busticketactivity.admin.getName
 import com.example.busticketactivity.dataclass.InfoTiket
 import com.example.busticketactivity.dataclass.ManagerGetData
 import com.example.busticketactivity.firebase.FireBaseRepo
 import com.google.gson.Gson
 import com.google.zxing.WriterException
+import kotlinx.android.synthetic.main.activity_detail_driver_t_icket.*
 import kotlinx.android.synthetic.main.activity_detail_tiket.*
+import kotlinx.android.synthetic.main.activity_detail_tiket.btn_cancel
+import kotlinx.android.synthetic.main.activity_detail_tiket.tv_jam_berangkat
+import kotlinx.android.synthetic.main.activity_detail_tiket.tv_nama_text
+import kotlinx.android.synthetic.main.activity_detail_tiket.tv_nomor_kursi
+import kotlinx.android.synthetic.main.activity_detail_tiket.tv_tanggal_pesan_berangkat
+import kotlinx.android.synthetic.main.activity_detail_tiket.tv_tanggal_text
+import kotlinx.android.synthetic.main.activity_detail_tiket.tv_terminal_text
+import kotlinx.android.synthetic.main.activity_detail_tiket.tv_tipe_bus_text
+import kotlinx.android.synthetic.main.activity_detail_tiket.tv_title
 import java.text.SimpleDateFormat
 import java.util.*
 
 
 class DetailTiketActivity : AppCompatActivity(), View.OnClickListener {
     val gson=Gson()
+    private var nama=""
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_detail_tiket)
@@ -32,15 +44,23 @@ class DetailTiketActivity : AppCompatActivity(), View.OnClickListener {
         val infoTiket = intent.getSerializableExtra("dataTiketPembayaran") as InfoTiket
         tv_jam_berangkat.text = infoTiket.pergi
         tv_nama_text.text = infoTiket.nama
-        tv_terminal.text = infoTiket.terminal
+        tv_terminal_text.text = infoTiket.terminal
         tv_nomor_kursi.text = infoTiket.nomorKursi
         tv_tanggal_pesan_berangkat.text=infoTiket.tanggalBeli
         tv_tanggal_text.text=infoTiket.tanggal
         tv_tipe_bus_text.text = infoTiket.type
+        FireBaseRepo().getUserNama(infoTiket.email).addOnCompleteListener {
+            if(it.isSuccessful){
+                val hasil=it.result!!.toObjects(getName::class.java)
+                nama=hasil[0].nama
+                tv_title.text=hasil!![0].nama
+            }
+        }
     }
 
     private fun getData() {
         val infoTiket = intent.getSerializableExtra("dataTiketPembayaran") as InfoTiket
+        infoTiket.namaUser=nama
         val gsontojson=gson.toJson(infoTiket,infoTiket::class.java)
         gsontojson
         val qrCode = QRGEncoder(gsontojson, null, QRGContents.Type.TEXT, 250)
@@ -73,14 +93,14 @@ class DetailTiketActivity : AppCompatActivity(), View.OnClickListener {
         val time = infoTiket.pergi.take(5)
         val datePergi = infoTiket.pergi.takeLast(11)
         val sdf = SimpleDateFormat("HHmm", Locale.getDefault())
-        val date = SimpleDateFormat("dd MMM yyyy", Locale.getDefault())
+        val date = SimpleDateFormat("dd-M-yyyy", Locale.getDefault())
         val dated = date.format(Date()).toString()
         val currentDateandTime: String = sdf.format(Date())
         val timeremove = time.replace(":", "")
-//        if (dated == datePergi) {
-//            if (timeremove.toInt() > currentDateandTime.toInt()) {
-//                val waktusisa = timeremove.toInt()
-//                if ((waktusisa - 100) >= currentDateandTime.toInt()) {
+        if (dated == infoTiket.tanggal) {
+            if (timeremove.toInt() > currentDateandTime.toInt()) {
+                val waktusisa = timeremove.toInt()
+                if ((waktusisa - 100) >= currentDateandTime.toInt()) {
                     FireBaseRepo().canceltiket(infoTiket.id,infoTiket.tanggal, infoTiket.nomorKursi)
                     FireBaseRepo().deletetiket(infoTiket.email).addOnCompleteListener {
                         if (it.isSuccessful) {
@@ -103,27 +123,27 @@ class DetailTiketActivity : AppCompatActivity(), View.OnClickListener {
                             }
                         }
                     }
-//                }
-    //                else {
-//                    Toast.makeText(
-//                        this,
-//                        "Waktu untuk pembatalan kurang dari 1 jam",
-//                        Toast.LENGTH_SHORT
-//                    ).show()
-//
-//                }
-//            } else {
-//                Toast.makeText(this, "Waktu untuk pembatalan kurang dari 1 jam", Toast.LENGTH_SHORT)
-//                    .show()
-//                Toast.makeText(this, "Waktu untuk pembatalan kurang dari 1 jam", Toast.LENGTH_SHORT)
-//                    .show()
-//            }
-//        } else {
-//            Toast.makeText(
-//                this,
-//                "Tanggal berbeda dengan keberangakatan tiket tidak bisa di cancel",
-//                Toast.LENGTH_SHORT
-//            ).show()
-//        }
+                }
+                    else {
+                    Toast.makeText(
+                        this,
+                        "Waktu untuk pembatalan kurang dari 1 jam",
+                        Toast.LENGTH_SHORT
+                    ).show()
+
+                }
+            } else {
+                Toast.makeText(this, "Waktu untuk pembatalan kurang dari 1 jam", Toast.LENGTH_SHORT)
+                    .show()
+                Toast.makeText(this, "Waktu untuk pembatalan kurang dari 1 jam", Toast.LENGTH_SHORT)
+                    .show()
+            }
+        } else {
+            Toast.makeText(
+                this,
+                "Tanggal berbeda dengan keberangakatan tiket tidak bisa di cancel",
+                Toast.LENGTH_SHORT
+            ).show()
+        }
     }
 }
